+++ rendering/frustum.h
@@ -10,7 +10,6 @@ http://mozilla.org/MPL/2.0/.
 #pragma once
 
 #include "utilities/Float3d.h"
-#include "utilities/dumb3d.h"
 
 inline std::vector<glm::vec4> const ndcfrustumshapepoints //
     {
